# server/src/spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SMOOTH API",
    "version": "1.0.0",
    "description": "Cameroon all-in-one platform API (EN/FR). Base URL: http://localhost:5000/api/v1. Money values are integers in XAF. Responses use the envelope { success, data, message?, meta? }.",
    "contact": { "name": "SMOOTH Team" }
  },
  "servers": [{ "url": "http://localhost:5000/api/v1", "description": "Local development" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/listings": {
      "get": {
        "tags": ["listings"],
        "summary": "Browse active listings (filters: q, city, status, category, bedrooms, minPrice, maxPrice, sort)",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["listings"],
        "summary": "Create listing (owner/agent)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/bookings": {
      "post": {
        "tags": ["bookings"],
        "summary": "Create booking (client)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["serviceType", "date", "price"], "properties": { "serviceType": { "type": "string" }, "providerId": { "type": "string" }, "listingId": { "type": "string" }, "date": { "type": "string", "format": "date" }, "price": { "type": "integer" }, "requirements": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Created, returns booking with ref" } }
      }
    },
...
    "/concierge": {
      "post": {
        "tags": ["concierge"],
        "summary": "Create concierge request (client)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["type", "details"], "properties": { "type": { "type": "string", "enum": ["errand", "delivery", "airport", "admin", "travel", "bill"] }, "details": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/jobs": {
      "get": {
        "tags": ["jobs"],
        "summary": "Browse active jobs (filters: q, category, city, type)",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["jobs"],
        "summary": "Post a job (employer)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["title", "category", "location", "description"], "properties": { "title": { "type": "string" }, "category": { "type": "string" }, "city": { "type": "string" }, "type": { "type": "string" }, "description": { "type": "string" }, "salary": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/payments/{ref}/receipt": {
      "get": { "tags": ["payments"], "summary": "Payment receipt (successful only)", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "ref", "in": "path", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/reviews": {
      "post": {
        "tags": ["reviews"],
        "summary": "Create review (client/agent/provider/employer)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["targetType", "targetId", "rating"], "properties": { "targetType": { "type": "string", "enum": ["listing", "agent", "provider", "employer"] }, "targetId": { "type": "string" }, "rating": { "type": "integer", "minimum": 1, "maximum": 5 }, "comment": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
...
    "/notifications": {
      "get": { "tags": ["notifications"], "summary": "My notifications + unread count", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/tickets": {
      "post": {
        "tags": ["tickets"],
        "summary": "Create support ticket",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["subject", "message"], "properties": { "category": { "type": "string", "enum": ["complaint", "question", "report_listing", "report_provider", "refund", "other"] }, "subject": { "type": "string" }, "message": { "type": "string" } } } } }
        },
        "responses": { "201": { "description": "Created, returns ticket ref (TCK-...)" } }
      }
    },
...
    "/uploads": {
      "post": {
        "tags": ["uploads"],
        "summary": "Upload file (multipart, field 'file'). max 10MB. Accepts jpg/png/webp/gif/pdf/heic",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } }
        },
        "responses": { "200": { "description": "OK, returns { url, provider }" } }
      }
    },
    "/admin/users": {
      "get": { "tags": ["admin"], "summary": "List users", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/admin/listings": {
      "get": { "tags": ["admin"], "summary": "All listings (any state)", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/admin/reviews": {
      "get": { "tags": ["admin"], "summary": "All reviews incl. hidden", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/admin/tickets": {
      "get": { "tags": ["admin"], "summary": "All support tickets", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
    },
...
  },
...
}